http: Add target schemes: dir, static, redirect
This patch adds to new target schemes for HTTP reverse proxying: dir,
static, and redirect.

The "dir" scheme will make gofer serve a directory directly, and
"static" will do the same but for just a file.

"redirect" will do an HTTP temporary redirection.

This can be convenient to quickly export static files, and redirect from
http to https without the need for an additional server.