http: Implement statusWriter.ReadFrom to enable sendfile
Go can transparently use the sendfile syscall to serve files, which can
be a large performance improvement; however this only happens if,
amongst other things, the response writer has a ReadFrom method.

This patch implements that ReadFrom method in statusWriter so it works
with sendfile, which yields a ~50% performance improvement for common
conditions such as file serving.