test: Add performance tests
This patch adds some performance tests, which do simple stress test
using the "wrk" benchmarking tool of gofer and nginx, and graph the
results.

These are not good tests; they are added only to help identify
regressions and get a rough idea of performance characteristics, but
they shouldn't be taken too seriously.