test/perf: Adjust parameters for more consistent results
The performance tests are useful to get a general sense of performance
in comparison with nginx, but they're meant to run on a single machine,
which can make them very noisy, especially on machines with few CPUs.

This patch does some minor adjustments to them, to make results more
consistent and less dependent on kernel scheduling decisions and
processor layout.

As a downside, the results become less representative in the total
requests per second, and won't highlight concurrency-related
bottlenecks. But it makes the tests much more useful for profiling and
catching large regressions.

The changes are:

- Run wrk with 1 thread, 1 request at a time. This improves consistency
  of measurement and drastically reduces measurement noise due to kernel
  decisions on machines with few CPUs.
- Enable sendfile on nginx, which gives it a significant latency
  improvement on large files.
- Cap Go runtime to 2 CPUs. This is to match nginx config, which also
  limits it to 2 CPUs.