trace: Only call log.Log if it's going to be useful
Calling log.Log is not necessary most of the time, since it is only
useful when running with debug verbosity. However, the string formatting
is a small but not insignificant cost when serving small files.

This patch makes the log.Log call conditional to the log level, to
avoid formatting the string unnecessarily.