Reduce log.Fatalf calls to make code easier to test
log.Fatalf calls cause an abrupt exit of the process, and can't be
properly tested. Some of these are reasonable and useful, but most are
not needed, we can just log and return an error instead.

This patch changes several `log.Fatalf` calls to `return log.Errorf`, to
make code easier to test.