ci: Always check for latest Go version in govulncheck
Occasionally, we may get govulncheck failing because the test is not run
with the latest Go release. This often happens after a new Go security
release, for example.

To help reduce the chances of these somewhat false positives, make the
GitHub CI always check for the latest Go version when running
govulncheck.