http proxy: Use Rewrite-based logic, improve forwarding headers
This patch updates the HTTP Proxy to use Rewrite instead of Director,
which is safer and more standard compliant, and has some nice
side-effects like improving the forwarding headers and logic for CGI
targets.

It also allow us to remove a fair amount of wrapping code that is no
longer needed.

It has one external behaviour change: the outgoing Host header used to
be set to the incoming Host header, but is now set to the proxy url, and
the original one is now available in the X-Forwarded-Host and Forwarded
headers.