Support key log writer for debugging
This patch adds an option to make gofer write TLS key logs to a file in
HTTPS connections, which is useful when debugging TLS-encrypted
sessions.

This is insecure and should only used for debugging purposes, so it is
not documented for now, and it is not expected to be a stable part of
the configuration.