reqlog: Use "-" when HTTP.Request.Host is empty
Today, we print HTTP.Request.Host as-is. So if it is empty, we just skip
the field in the logs.

A log line with Host value: `... HTTP/1.1 example.com GET ...`.

And with Host empty: `... HTTP/1.1  GET ...` (note the double spaces).

This makes the logs more difficult to read and parse.

To fix this, this patch makes gofer log `-` when the host is empty.

After this change, the a log line will look like:
`... HTTP/1.1 - GET ...` (note the `-` in the host field).